requirements.txt
beautifulsoup4==4.11.1
blurhash==1.1.4
certifi==2022.12.7
charset-normalizer==2.1.1
decorator==5.1.1
idna==3.4
Mastodon.py==1.8.0
mysql-connector-python==8.0.31
protobuf==3.20.1
python-dateutil==2.8.2
python-magic==0.4.27
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.13
